requirements.txt
fastapi
uvicorn
pydantic
huggingface_hub
pycatia
python-dotenv
requests
torch
transformers
accelerate
